Rename install to update 🔠
