+++ src/run/update/update-npm-scripts.js
